Refactored faidx indexing
Instead of separating the indexing for fasta and fastq files into
separate reader objects, they were combined into a single indexer object
that lives in the fai io module.  This removes the need to use Box for
building the reader in the cmd module.  This also removes a lot of
uncessary duplicate code and should be more simple to maintain.